refactor: button 아이콘 폴더로 이동
