+++ frontend/src/components/BreedSearch.tsx
@@ -4,7 +4,7 @@ import TopBack from '@/assets/icons/ic-top-back.svg';
 import Search from '@/assets/icons/ic-search.svg';
 import { DogRegInfo } from '@/pages/Join';
 import { useBreed } from '@/hooks/useBreed';
-import DeleteBtn from '@/assets/icons/btn-delete.svg';
+import DeleteBtn from '@/assets/buttons/btn-delete.svg';
 interface Props {
     isOpen?: boolean;
     setIsOpen?: (state: boolean) => void;
